Move tests next to source folder
